Refactor redundant code
